[feature] 避免get方法传数组

+++ update-main/src/main/java/com/thinking/update/main/dao/AppDao.java
@@ -74,7 +74,7 @@ public interface AppDao{
 	 * @return
 	 */
 	List<App> getAppForPageByDeviceIdsAndStateList(@Param("states") List<Integer> states,
-												   @Param("deviceIds") List<Integer> deviceIds);
+												   @Param("deviceIds") List<Long> deviceIds);
 	/**
 	 * 通过App的id获得App对象
 	 * @param id
